requirements.txt
tensorflow
numpy
pillow
scikit-learn
streamlit
matplotlib
